+++ src/components/super/i-static-page/modules/ss-helpers/tags.js
@@ -15,7 +15,6 @@ const
 
 const
 	fs = require('fs-extra'),
-	delay = require('delay'),
 	buble = require('buble');
 
 const
@@ -120,9 +119,9 @@ function getScriptDecl(lib, body = '') {
 	}
 
 	if (isInline && !body) {
-		return (async () => {
-			while (!fs.existsSync(lib.src)) {
-				await delay(500);
+		return (() => {
+			if (!fs.existsSync(lib.src)) {
+				throw new Error(`The asset ${lib.src} cannot be found`);
 			}
 
 			const
@@ -244,9 +243,9 @@ function getStyleDecl(lib, body = '') {
 		attrs = normalizeAttrs(attrsObj, lib.js);
 
 	if (isInline && !body) {
-		return (async () => {
-			while (!fs.existsSync(lib.src)) {
-				await delay(500);
+		return (() => {
+			if (!fs.existsSync(lib.src)) {
+				throw new Error(`The asset ${lib.src} cannot be found`);
 			}
 
 			if (lib.js) {